requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
fastapi==0.117.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-documentai==3.6.0
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.76
langchain-openai==0.3.33
langsmith==0.4.29
lxml==6.0.2
MarkupSafe==3.0.2
nodeenv==1.9.1
numpy==2.3.3
openai==1.108.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
prisma==0.15.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
zstandard==0.25.0
